mostly fixed, it happens once if you refresh the page
