+++ webui/react/src/pages/TrialDetails/TrialDetailsProfiles.tsx
@@ -5,6 +5,8 @@ import ProfilesEnabled from 'pages/TrialDetails/Profiles/ProfilesEnabled';
 import ProfilesFiltersProvider from 'pages/TrialDetails/Profiles/ProfilesFiltersProvider';
 import ProfilesNotEnabled from 'pages/TrialDetails/Profiles/ProfilesNotEnabled';
 import { ExperimentBase, TrialDetails } from 'types';
+import useScroll from 'hooks/useScroll';
+
 
 import css from './TrialDetailsProfiles.module.scss';
 
@@ -37,8 +39,23 @@ const Profiler: React.FC<Props> = ({ experiment, trial }: Props) => {
 };
 
 const TrialDetailsProfiles: React.FC<Props> = (props: Props) => {
+  const containerRef = useRef<HTMLDivElement>(null);
+  const scroll = useScroll(containerRef);
+  const scrollTop = useRef(0);
+  
+    /*
+   * Preserve and restore scroll position upon re-render.
+   */
+  
+  useEffect(() => {
+    if (containerRef.current && scroll.scrollTop === 0 && scrollTop.current !== 0) {
+      containerRef.current.scrollTop = scrollTop.current;
+    } else {
+      scrollTop.current = scroll.scrollTop;
+    }
+  }/* , [ scroll ] */);
   return (
-    <div className={css.base}>
+    <div ref={containerRef} className={css.base}>
       <Profiler {...props} />
     </div>
   );
